redux-thunk added, proccessing of files moved to action folder
