+++ package.json
@@ -21,6 +21,7 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.0",
     "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
     "styled-components": "^4.4.1",
     "styled-normalize": "^8.0.6"
   },
