+++ src/components/Wrappers/StoreProvider.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { createStore } from 'redux'
+import { createStore, applyMiddleware } from 'redux'
 import { Provider } from 'react-redux'
-
+import thunk from 'redux-thunk'
 import rootReducer from '../../reducers/index'
 
-const store = createStore( rootReducer )
+const store = createStore( rootReducer, applyMiddleware(thunk) )
 
 const StoreProvider = ({children}) => (
   <Provider store={store}>
